Enable boot from xHCI with u-boot v2018.05
Update u-boot to v2018.05 with fixes to xHCI hub handling and the
xhci-rcar driver. The u-boot dtb now activates the xHCI controller and
the default bootcmd initializes the on-board hub. This allows booting
from a USB storage device on the CETiBOX external USB ports.

The linux R-Car xHCI driver is now compiled into the kernel to allow
mounting the rootfs from a USB storage device on the external USB ports.
The required firmware is supplied via a firmware-only initramfs (init is
loaded directly from the 'real' rootfs).