u-boot-fw-utils: Use negative offset to support different mmcblk0boot0 sizes
Configure the userspace fw_*env tools with a negative environment offset,
causing them to look for the environment at the specified offset from the
end of the device. This matches the behaviour of u-boot itself and allows
userspace environment access to work with any mmcblk0boot0 size.
Fixes CBX3IST-158.