Workaround for inconsistent MOST startup behaviour.
The MOST async channel is not set up correctly during every boot, if it
is configured through the udev rules in most-driver.rules. There seems
to be a race condition in the driver, leading to incomplete MOST
configuration when the async tx channel is initialized before the async
rx channel.

Moved the complete initialization to a separate systemd path unit which
waits for the corresponding sysfs nodes to pop up before an
configuration attempt is made. This workaround is stable so far,
however we need to investigate the real cause for this issue and fix it.