Wait for all most/mostcore/devices/mdev? to appear before starting most driver
In ffeff2b2, the MOST startup scripting was changed to work around the
driver creating the mdev's in inconsistent order. The systemd path unit
used there contained multiple PathExists= to detect when all mdev[012] have
been created. But this triggers when any of the paths is created.

To achieve the desired behaviour of triggering when _all_ of the paths
exits, tag these paths in udev and make most-driver-setup.service
depend on the sys-devices-..mdev[012].device nodes. The path unit is
removed. Now, the MOST startup script is started when all of the
mdev[012] have been created.