Prevent early stage bootloaders from driving EN pin of LIN transceiver.
The enable pin of the LIN transceiver (SH65HVDA195) must not be actively
driven by the R-Car if the transceiver isn't used for LIN communication.
Otherwise the transceiver will generate a wake up pulse on the LIN line,
waking up other devices as soon as the CETiBOX is turned on.

This issue is resolved by setting GPSR2_7 to 0, effectively letting the
line floating. A weak pulldown in the transceiver makes sure that the
line is in a deterministic state.

Note:
Also eliminated a few spacely-tabbed lines in the bb file.