linux-cetibox: Fix firmware fetching
The URLs used to fetch files from linux-firmware did not specify a tag
and so always fetched the latest revision. Since sometimes firmware
files are updated in place, this caused the checksum verification to
fail.

Fix this by specifying a tagged revision for fetching firmwares so that
we always fetch the same version. Also update the firmware checksums to
sha256.

Fixes CBX3IST-289.