sja1105: Make switch configuration units RemainAfterExit
As the switch configuration persists in the switch once configured, even
when the configuration utility has exited, it makes sense to set
RemainAfterExit to yes. This allows units which are possibly started
later (such as the gPTP bridge) to depend on the switches being
configured without triggering an unnecessary reconfiguration of the
switches.